test: split sk
