+++ packages/core/src/sheets/util.ts
@@ -24,6 +24,11 @@ import { Rectangle } from '../shared';
 import { type CellValueType, HorizontalAlign, type TextDirection, VerticalAlign, WrapStrategy } from '../types/enum';
 import { CustomRangeType, type IDocumentData, type IPaddingData, type IStyleBase, type IStyleData, type ITextRotation, type ITextStyle } from '../types/interfaces';
 
+export interface IFontLocale {
+    fontList: string[];
+    defaultFontSize: number;
+}
+
 export const isRangesEqual = (oldRanges: IRange[], ranges: IRange[]): boolean => {
     return ranges.length === oldRanges.length && !oldRanges.some((oldRange) => ranges.some((range) => !Rectangle.equals(range, oldRange)));
 };
