+++ packages/engine-render/tsconfig.json
@@ -5,5 +5,5 @@
         "outDir": "lib/types"
     },
     "references": [{ "path": "./tsconfig.node.json" }],
-    "include": ["src", "types"]
+    "include": ["src", "types", "../core/src/skeleton.ts"]
 }
