+++ packages/sheets-ui/src/controllers/utils/component-tools.ts
@@ -127,7 +127,7 @@ export function getCoordByOffset(
 
     const { row, column } = moveActualSelection;
 
-    const startCell = skeleton.getNoMergeCellPositionByIndex(row, column);
+    const startCell = skeleton.getNoMergeCellWithCoordByIndex(row, column);
 
     const { startX, startY, endX, endY } = startCell;
